rename ExpansionContext to InfoContext
