+++ packages/core/contexts/index.ts
@@ -2,9 +2,9 @@ export * from './AboutContext'
 export * from './AddonContext'
 export * from './BoardContext'
 export * from './ClientContext'
-export * from './ExpansionContext'
 export * from './FitContext'
 export * from './IllustContext'
+export * from './InfoContext'
 export * from './LoggingContext'
 export * from './PaddingContext'
 export * from './PickerContext'
